🚚 move: SettingPage components/
