+++ app/src/pages/SettingPage/components/AccountSection.tsx
@@ -5,14 +5,14 @@ import {
   Spacer,
   VStack,
 } from '@components/common';
-import styled from '@emotion/styled';
+import { NeumorphismSection } from '@styles/custom/NeumorphismSection';
 import { Mobile, TabletAndAbove } from '@utils/responsive/Device';
 import { DeleteAccountButton } from './DeleteAccountButton';
 import { LogoutButton } from './LogoutButton';
 
 export const AccountSection = () => {
   return (
-    <Layout>
+    <NeumorphismSection>
       <TabletAndAbove>
         <HStack>
           <H2BoldText>계정 관리</H2BoldText>
@@ -33,19 +33,6 @@ export const AccountSection = () => {
           </HStack>
         </VStack>
       </Mobile>
-    </Layout>
+    </NeumorphismSection>
   );
 };
-
-const Layout = styled.div`
-  width: 100%;
-  padding: 3rem;
-  border-radius: ${({ theme }) => theme.radius.md};
-  box-shadow: 10px 10px 10px #e8e8e8, -10px -10px 10px #ffffff;
-
-  transition: all 0.3s;
-  &:hover {
-    transform: scale(100.5%);
-    box-shadow: 10px 10px 10px #d2d2d2, -10px -10px 10px #ffffff;
-  }
-`;
