+++ app/src/pages/SettingPage/components/DeleteAccountButton.tsx
